backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
requests>=2.31.0 
pydantic-settings==2.0.0
psycopg2-binary==2.9.10
internetarchive